Removed a cast warning from StrokeArc
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@1968 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/app/proto7/ScreenDriver.cpp
@@ -785,7 +785,7 @@ void ScreenDriver::StrokeArc(BRect r, float angle, float span, LayerData *d, int
 	int startQuad, endQuad;
 	bool useQuad1, useQuad2, useQuad3, useQuad4;
 	bool shortspan = false;
-	int thick = d->pensize;
+	int thick = (int)d->pensize;
 
 	// Watch out for bozos giving us whacko spans
 	if ( (span >= 360) || (span <= -360) )
